issue-637: DevLoginController — drop DbContext, route through services
DevLoginController violated design-rules §2a (Web-layer controllers
must not inject DbContext) and §2c (cross-section writes flow through
the owning section's service). Same pattern fixed in PR #403 for
AccountController and AccountProvisioningService.

Persona seeding logic (User + Profile + UserEmail + auxiliary teams /
camps / roles / contact fields) is extracted into DevPersonaSeeder
under Web/Infrastructure, mirroring the existing
DevelopmentDashboardSeeder shape. The seeder is dev-only (registered
behind !IsProduction()).

In the seeder, cross-section writes for the canonical persona surface
flow through the owning services:
- User → UserManager<User>.CreateAsync
- UserEmail → IUserEmailService.AddVerifiedEmailAsync
- Profile → IProfileService.SaveProfileAsync (then a post-save direct
  IsApproved/ConsentCheckStatus override to preserve the
  audit-log-unchanged dev-seed behavior; routing through
  ApproveVolunteerAsync / RecordConsentCheckAsync would have added
  audit entries the controller never produced).

The auxiliary writes (TeamMembers to system teams, RoleAssignments,
ContactFields, dev-test Department / SubTeam, dev barrio Camp /
CampSeason / CampLead, dev city-planning Team) keep direct DbContext
writes inside the seeder — these are dev-only fixtures with no
production service surface for "create a fresh row with a
deterministic id", same shape as DevelopmentDashboardSeeder's
direct-write to Shifts-owned tables.

Net result: DevLoginController's constructor takes no HumansDbContext;
no _db.X.Add(...) / SaveChangesAsync() / Remove(...) calls remain in
the controller file.

Closes nobodies-collective/Humans#637
